Relocate TaskParametersConfigurator to shared components and split ManageVariants forms
