+++ apps/dashboard/src/components/TaskParametersConfigurator/TaskParametersConfigurator.vue
