+++ apps/dashboard/src/components/TaskParametersConfigurator/TaskParametersConfiguratorRow.vue
